From single-key to multi-key
Previously, proposers and acceptors operated on an implicit single key.
Now, key is taken as a parameter to the client-facing Propose function,
as well as the internal prepare and accept phases.

This requires no change in state in the proposer, which simply forwards
requests along as before, taking ballot numbers for all keys from its
single source of truth. This should work fine, because ballot number
counters merely must be monotonically increasing: it's OK if some
numbers are "skipped" on a per-key basis.

This requires a small change in state in the acceptor, namely it keeps a
map of keys to (Promise, Accepted, Value) tuples. Each prepare and
accept request applies to a single key/tuple pair. rystsov assures me
this is valid :)

A change to the configuration change test is required. A new acceptor
won't receive the value of a key until it undergoes a specific read.
This makes sense but is a change from before, where the single key was
implicit in the conf change and thus could be checked immediately.